safari-600.7-branch/LayoutTests/http/tests/security/no-javascript-refresh-static-percent-escaped.html

<!DOCTYPE html>
<head>
<meta http-equiv='refresh' content='0;URL=%6A%61%76%61%73%63%72%69%70%74:window.top.location="about:blank"'/>
</head>
<body>
<script>
if (window.testRunner)
    testRunner.dumpAsText();
</script>
PASS: This is the text that appears in place of a failed refresh.
</body>
