LayoutTests/http/tests/security/no-javascript-refresh-static-percent-escaped-expected.txt

PASS: This is the text that appears in place of a failed refresh.
